# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Headnet control-plane API
  version: "1.0.0"
  summary: Self-hosted mesh VPN control plane
  description: |
    The HTTP API of the Headnet control plane.

    **This document describes only what is implemented.** Endpoints for
    features that have not been built are deliberately absent rather than
    present-and-stubbed: a specification that lists endpoints returning
    fabricated data is worse than one that is honestly short.

    Planned endpoints are listed in
    [docs/ROADMAP.md](https://github.com/CauaMora1s/Headnet/blob/main/docs/ROADMAP.md)
    and will be added here as they land.

    ## Versioning

    The API is versioned in its path (`/api/v1`). The version changes only for
    a breaking change; additive changes — new fields, new endpoints — happen
    in place.

    **Clients must ignore fields they do not recognise.** That requirement is
    what makes additive evolution possible.

    Separately from the API version, a *protocol* version governs the
    client/server contract as a whole. See `/api/v1/protocol/negotiate`.

    ## Errors

    Every non-2xx response carries the same envelope, so a client parses one
    shape:

    ```json
    {
      "error": {
        "code": "bad_request",
        "message": "protocol_version must be a positive integer",
        "details": { "field": "protocol_version" },
        "request_id": "req_4KQZ9V3B2N7XR8T0YCFHJM"
      }
    }
    ```

...

    ## Rate limiting

...

...
    not trusted. See
...

...

...
servers:
...
paths:
...
  /api/v1/auth/bootstrap:
    post:
      tags: [Authentication]
      summary: Create the first administrator
      description: |
        Creates the first account on a fresh installation and signs it in. The
        account is necessarily an administrator: there is nobody else to grant
        it the role later.

        **This endpoint is open until it is claimed.** That is a deliberate
        trade-off in favour of the "deploy, open the web UI, create an account"
        flow, and it has a real cost: a publicly reachable server that nobody
        has claimed can be claimed by whoever finds it first. Claim yours
        immediately, and do not expose the server publicly until you have. The
        server logs a warning on every start until it is claimed.

        The claim is exactly-once and atomic, so concurrent requests cannot
        both succeed. A rejected account — a password that is too short, say —
        does not consume the claim.
      operationId: bootstrap
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BootstrapRequest" }
      responses:
        "201":
          description: The administrator was created and signed in.
          headers:
            Set-Cookie:
              schema: { type: string }
              description: Sets `headnet_session` and `headnet_csrf`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "400":
          description: The email address or password was rejected.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "409":
          description: The installation already has an administrator.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "429": { $ref: "#/components/responses/RateLimited" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: headnet_session
      description: |
        Server-side sessions in a `Secure`, `HttpOnly`, `SameSite=Lax` cookie.
        `Secure` is set whenever the deployment's base URL is HTTPS.

        Sessions are deliberately not stateless tokens: when a device is
        stolen, an administrator must be able to revoke access immediately,
        and a self-contained token cannot be revoked before it expires.

        State-changing requests must additionally carry the `X-CSRF-Token`
        header. See the Authentication tag.
  responses:
    RateLimited:
      description: |
        The client exceeded its request budget. `Retry-After` gives the number
        of seconds to wait.
      headers:
        Retry-After:
          schema: { type: integer }
          description: Seconds to wait before retrying.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            limited:
              value:
                error:
                  code: rate_limited
                  message: too many requests; slow down and try again shortly
                  details:
                    retry_after_seconds: "12"
...
  schemas:
...
    ErrorCode:
      type: string
      description: |
        A stable, machine-readable classification. Codes may be added but are
        never renamed or repurposed, so branching on them is safe.
      enum:
        - bad_request
        - unauthorized
        - forbidden
        - not_found
        - conflict
        - payload_too_large
        - rate_limited
        - unsupported_protocol_version
        - not_implemented
        - unavailable
        - internal
    Error:
      type: object
      required: [code, message]
      properties:
        code: { $ref: "#/components/schemas/ErrorCode" }
        message:
          type: string
          description: |
            A human-readable explanation, safe to show an operator. It never
            contains secrets, key material or internal stack detail — and it
            may change at any time, so do not parse it.
        details:
          type: object
          additionalProperties: { type: string }
          description: |
            Structured, code-specific context: the field that failed
            validation, the roadmap phase behind a `not_implemented`, and so
            on.
        request_id:
          type: string
          description: |
            Correlates this response with the server log. Also returned in the
            `X-Request-Id` header. Quote it in a bug report.
          examples: [req_4KQZ9V3B2N7XR8T0YCFHJM]
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { $ref: "#/components/schemas/Error" }
...
    BootstrapRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password:
          type: string
          format: password
          description: |
            At least `auth.min_password_length` characters, and never fewer
            than 12 whatever the configuration says. Length is asked for rather
            than composition rules, which mostly produce `Password1!`.
        display_name: { type: string }
    User:
      type: object
      required: [id, email, role, provider, created_at]
      description: |
        An account. There is deliberately no password field of any kind — not
        the hash, not a placeholder. A schema with somewhere to put one is a
        schema that will eventually carry one.
      properties:
        id:
          type: string
          examples: ["usr_4KQZ9V3B2N7XR8T0YCFHJM"]
        email: { type: string, format: email }
        display_name: { type: string }
        role: { type: string, enum: [admin, member] }
        provider: { type: string, enum: [local, oidc] }
        created_at: { type: string, format: date-time }
        last_login_at: { type: string, format: date-time }
    Session:
      type: object
      required: [user, csrf_token, expires_at]
      properties:
        user: { $ref: "#/components/schemas/User" }
        csrf_token:
          type: string
          description: |
            Echo this in `X-CSRF-Token` on every state-changing request. It is
            also set as the readable `headnet_csrf` cookie.

            Empty on `GET /api/v1/auth/session`: only a hash is stored, so it
            cannot be handed back after issue.
        expires_at:
          type: string
          format: date-time